requirements.txt
fastapi[standard]
openai-agents
requests
openinference-instrumentation-openai-agents
opentelemetry-sdk~=1.30.0
opentelemetry-exporter-otlp-proto-grpc~=1.30.0
opentelemetry-distro[otlp]~=0.51b0
